Tweaks for large files

+++ bench.sh
@@ -5,15 +5,18 @@ IFS=$'\n\t'
 
 test_duration=30
 test_concurrency=${VUS:-200}
-tests=(synthetic html proxy)
-proxies=(caddy nginx)
-suffix=${SUFFIX:-''}
-if [[ -z "${suffix}" ]]
-then
-    append=''
-else
-    append=" (${suffix})"
-fi
+
+declare -A proxies
+proxies[caddy]=Caddyfile
+proxies[nginx]=nginx.conf
+
+declare -A tests
+tests[synthetic]=''
+tests[proxy]=''
+tests[html_small]='/index.html'
+tests[html_large]='/jquery-3.6.1.js'
+
+levels=(default optimized)
 
 function ssh_ { ssh -F ssh_config $@; }
 function scp_ { scp -F ssh_config $@; }
@@ -27,6 +30,9 @@ function config {
 }
 
 function run_tests {
+    #
+    # Build machines
+    #
     terraform apply -auto-approve
     dns=$(terraform output -json dns)
     sut_pri=$(echo $dns | jq -r '.[0][0]')
@@ -43,35 +49,45 @@ function run_tests {
         done
     done
 
-    config aws-bench $sut_pub
+    #
+    # Configure driver
+    #
     config aws-driver $dri_pub
-    ssh_ $sut_pub systemctl stop nginx caddy || true
-
     scp_ test.js $dri_pub:~/
 
-    for svc in ${proxies[@]}
+    for svc in ${!proxies[@]}
     do
-        for test_uri in ${tests[@]}
+        for opt in ${levels[@]}
         do
-            ssh_ $sut_pub systemctl start $svc
-            sleep 3
-            pid=$(ssh_ $sut_pub systemctl show $svc --property=MainPID --value)
-            ssh_ $sut_pub "psrecord $pid \
-                --include-children \
-                --interval 0.1 \
-                --duration $(( $test_duration + 15 )) \
-                --log $svc-$test_uri-$test_concurrency$suffix.txt" &
-            record=$!
-            ssh_ $dri_pub \
-                 "TEST_TARGET=http://${sut_pri}:8080/${test_uri} \
-                k6 run \
-                --vus $test_concurrency \
-                --duration ${test_duration}s \
-                test.js"
-            ssh_ $sut_pub systemctl stop $svc
-            scp_ $dri_pub:~/summary.json results/${svc}-${test_uri}-${test_concurrency}$suffix.json
-            wait $record
-            scp_ $sut_pub:~/${svc}-${test_uri}-${test_concurrency}${suffix}.txt results/
+            if [[ ${svc} == "caddy" && ${opt} == "optimized" ]]
+            then
+                continue
+            fi
+            for test_ in ${!tests[@]}
+            do
+                cp -f conf/$svc/$opt/$test_ ${proxies[$svc]}
+                config aws-bench $sut_pub
+
+                ssh_ $sut_pub systemctl start $svc
+                sleep 3
+                pid=$(ssh_ $sut_pub systemctl show $svc --property=MainPID --value)
+                ssh_ $sut_pub "psrecord $pid \
+                    --include-children \
+                    --interval 0.1 \
+                    --duration $(( $test_duration + 15 )) \
+                    --log $svc-$opt-$test_-$test_concurrency.txt" &
+                record=$!
+                ssh_ $dri_pub \
+                    "TEST_TARGET=http://${sut_pri}:8080${tests[$test_]} \
+                    k6 run \
+                    --vus $test_concurrency \
+                    --duration ${test_duration}s \
+                    test.js"
+                ssh_ $sut_pub systemctl stop $svc
+                scp_ $dri_pub:~/summary.json results/${svc}-${opt}-${test_}-${test_concurrency}.json
+                wait $record
+                scp_ $sut_pub:~/${svc}-${opt}-${test_}-${test_concurrency}.txt results/
+            done
         done
     done
 
@@ -79,48 +95,54 @@ function run_tests {
 }
 
 function postprocess_resources {
-    cleanup=()
-    for test_type in ${tests[@]}
+    results=()
+    for test_ in ${!tests[@]}
     do
-        for svc in ${proxies[@]}
+        for opt in ${levels[@]}
         do
-            raw=results/${svc}-${test_type}-${test_concurrency}${suffix}.txt
-            ready=results/${svc}-${test_type}-${test_concurrency}${suffix}-formatted.txt
-            cleanup+=($ready)
+            for svc in ${!proxies[@]}
+            do
+                if [[ ${svc} == "caddy" && ${opt} == "optimized" ]]
+                then
+                    continue
+                fi
+                raw=results/${svc}-${opt}-${test_}-${test_concurrency}.txt
 
-            sed 1d $raw \
-                | choose 0:2 \
-                | sed '1i Time "'$svc' CPU %" "'$svc' Memory"' \
-                      > $ready
+                sed 1d $raw \
+                    | choose 0:2 \
+                    | sed '1i Time "'$svc' '${test_/_/ }' '${opt}' CPU %" "'$svc' '${test_/_/ }' '${opt}' Memory"' \
+                    | sponge $raw
+                results+=(-e "${svc}_${opt}_${test_}='${raw}'")
+            done
         done
-
-        gnuplot \
-            -e "test_type='${test_type}'" \
-            -e "par='${test_concurrency}'" \
-            -e "caddy='results/caddy-$test_type-$test_concurrency$suffix-formatted.txt'" \
-            -e "nginx='results/nginx-$test_type-$test_concurrency$suffix-formatted.txt'" \
-            -e "append='${append}'" \
-            resources.gp \
-            > results/resources-${test_type}-${test_concurrency}c${suffix}.svg
     done
-    rm ${cleanup[@]}
+
+    gnuplot -e "par='${test_concurrency}'" ${results[@]} resources.gp \
+        > results/resources-${test_concurrency}c.svg
 }
 
 function postprocess_metrics {
     echo "test min median average p90 p95 max requests errors" > results/plot.txt
-    for test_type in ${tests[@]}
+    for test_ in ${!tests[@]}
     do
-        for svc in ${proxies[@]}
+        for opt in ${levels[@]}
         do
-            raw=results/${svc}-${test_type}-${test_concurrency}${suffix}.json
-            jq --arg var http_req_duration -r -f metric.jq $raw \
-                | xargs echo "${svc}-${test_type}" \
-                        >> results/plot.txt
+            for svc in ${!proxies[@]}
+            do
+                if [[ ${svc} == "caddy" && ${opt} == "optimized" ]]
+                then
+                    continue
+                fi
+                raw=results/${svc}-${opt}-${test_}-${test_concurrency}.json
+                jq --arg var http_req_duration -r -f metric.jq $raw \
+                    | xargs echo "${svc}-${opt}-${test_/_/-}" \
+                            >> results/plot.txt
+            done
         done
     done
 
     rs -Tc' ' < results/plot.txt | sponge results/plot.txt
-    cp results/{plot.txt,table-${test_concurrency}${suffix}.txt}
+    cp results/{plot.txt,table-${test_concurrency}}
     sed -n '1p;/requests/p' results/plot.txt > results/requests.txt
     sed -n '1p;/error/p' results/plot.txt > results/errors.txt
     sed -i '/requests/d;/error/d;/max/d;/min/d' results/plot.txt
@@ -130,8 +152,7 @@ function postprocess_metrics {
             -e "errors='results/errors.txt'" \
             -e "concurrency='$test_concurrency'" \
             -e "test_type='duration'" \
-            -e "append='${append}'" \
-            metrics.gp > results/metrics-duration-${test_concurrency}c${suffix}.svg
+            metrics.gp > results/metrics-duration-${test_concurrency}c.svg
     rm results/{errors,plot,requests}.txt
 }
